# Conflicts:
#	Assets/Scenes/LoginScene.unity
#	Assets/Scenes/MainScene.unity
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-01.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-02.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-03.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-04.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-05.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-06.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-07.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-08.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-09.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-10.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-11.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-12.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-13.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-14.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-15.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-16.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-17.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-18.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-19.asset
#	Assets/ScriptableObjects/Spawn/Chapter01/Stage01-20.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-01.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-02.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-03.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-04.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-05.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-06.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-07.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-08.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-09.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-10.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-11.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-12.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-13.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-14.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-15.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-16.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-17.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-18.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-19.asset
#	Assets/ScriptableObjects/Spawn/Chapter02/Stage02-20.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-01.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-02.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-03.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-04.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-05.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-06.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-07.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-08.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-09.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-10.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-11.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-12.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-13.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-14.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-15.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-16.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-17.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-18.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-19.asset
#	Assets/ScriptableObjects/Spawn/Chapter03/Stage03-20.asset
#	Assets/Scripts/Chapter/ChapterData.cs
#	Assets/Scripts/Chapter/Stage/StageData.cs
#	Assets/Scripts/Currency/CurrencyDrop.cs
#	Assets/Scripts/Datas/Server/ServerData.cs
#	Assets/Scripts/Datas/Server/ServerData_Users.cs
#	Assets/Scripts/Manager/ChapterManager.cs
#	Assets/Scripts/Manager/LoginManager.cs

+++ Assets/Scripts/Components/EnemyRewardComponent.cs
@@ -10,17 +10,17 @@ public class EnemyRewardComponent : MonoBehaviour, IReward
     /// <summary>
     /// Gold 양을 반환하는 프로퍼티
     /// </summary>
-    public uint Gold { get; private set; }
+    public ulong Gold { get; private set; }
 
     /// <summary>
     /// Gem 양을 반환하는 프로퍼티
     /// </summary>
-    public uint Gem { get; private set; }
+    public ulong Gem { get; private set; }
 
     /// <summary>
     /// Exp 양을 반환하는 프로퍼티
     /// </summary>
-    public uint Exp { get; private set; }
+    public ulong Exp { get; private set; }
 
     /// <summary>
     /// 캐릭터의 고유 ID를 가져오는 인터페이스
